Separate durable append error classification
Durable conversation recovery had append error parsing and classification embedded in the broad durable module. Moving that boundary into a focused module makes the recovery path easier to reason about while preserving the public durable exports used by existing callers.

Constraint: Public imports from durable.ts and veryfront/agent must remain stable.

Rejected: Move fetch-backed cursor recovery in the same PR | it depends on getConversationRun and would require a larger module-boundary change.

Confidence: high

Scope-risk: narrow

Tested: red-first deno test --no-check --allow-all src/agent/conversation/durable-append-errors.test.ts failed with missing helper module

Tested: deno test --no-check --allow-all src/agent/conversation/durable-append-errors.test.ts src/agent/conversation/durable.test.ts

Tested: deno test --no-check --allow-all src/agent/conversation

Tested: deno check src/agent/conversation/durable.ts src/agent/conversation/durable-append-errors.ts src/agent/conversation/durable-append-errors.test.ts src/agent/index.ts

Tested: deno lint src/agent/conversation/durable.ts src/agent/conversation/durable-append-errors.ts src/agent/conversation/durable-append-errors.test.ts

Tested: deno fmt --check src/agent/conversation/durable.ts src/agent/conversation/durable-append-errors.ts src/agent/conversation/durable-append-errors.test.ts

Tested: git diff --check
